added IIndexArtifactMap interface in order to support permanent resourceIds
git-svn-id: https://hyracks.googlecode.com/svn/branches/hyracks_lsm_tree@1591 123451ca-8445-de46-9d55-352943316053

+++ hyracks-storage-am-common/src/main/java/edu/uci/ics/hyracks/storage/am/common/dataflow/IndexRegistry.java
@@ -19,17 +19,17 @@ import java.util.HashMap;
 
 public class IndexRegistry<IndexType> {
 
-    private HashMap<Integer, IndexType> map = new HashMap<Integer, IndexType>();
+    private HashMap<Long, IndexType> map = new HashMap<Long, IndexType>();
 
-    public IndexType get(int indexId) {
+    public IndexType get(long indexId) {
         return map.get(indexId);
     }
 
-    public void register(int indexId, IndexType index) {
+    public void register(long indexId, IndexType index) {
         map.put(indexId, index);
     }
 
-    public void unregister(int indexId) {
+    public void unregister(long indexId) {
         map.remove(indexId);
     }
